+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Breaking
 
 Changed
   * AZ-572: Revamp examples and improve CI
+  * AZ-302: Bump Log Analytics version to latest and allow override
 
 # v4.2.1 - 2021-08-27